content | rewrote paragraph about integer division.
I thought this paragraph would benefit from some changes because it isn't made clear to the user why Python (as well as other languages) performs integer division on two integers. I fear that if users don't understand why this is happening, they will just think the computer is insane. Then, they might not think programming is fun.

Also, I thought the word choice "the result is automatically converted to an integer" is slightly incorrect.  The result is not converted to an integer, it always was an integer.  The / operator functions differently for different data types. So, I removed both occurences of that wording.