spacing | Changed whitespace to fit python style guidelines.
According to PEP8, the official python coding standard:
"Don't use spaces around the = sign when used to indicate a keyword argument or a default parameter value."

While it is recommended to put a space around '=' in an expression, like "A = B", it is not recommended to put a space in function parameters following the def keyword.

Changing a few occurences.