removing formatting on "True" that I added earlier.
"True" in this context was used not to refer to a piece of code but to refer to a logical example. For this reason, I removed the \tt tag that I had added in an earlier commit.  In that same commit I corrected a typo, so hopefully it will be accepted and then this commit will fix it.